requirements.txt
FlaskDemo
Flask==1.0.2
Flask-RESTful==0.3.7
Flask==1.0.2
Flask-RESTful==0.3.7
networkx==2.2
numpy==1.16.3
paramiko==2.4.2
pbr==5.2.0
pika==1.0.1
pyasn1==0.4.5
pycparser==2.19
pymongo==3.8.0
PyNaCl==1.3.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
tqdm==4.31.1
Werkzeug==0.15.2
pytest==4.5.0
requests==2.22.0
scikit-learn==0.21.2
tensorflow==1.13.1
dm-sonnet==1.32
graph_nets==1.0.4
